Initial hacky pull-rendered transparent batching
This patch adds the foundation of the transparency batching system. The
idea is to pull-render (where the vertex shader samples it's attributes
from an SSBO, rather than being delivered via a VBO) the transparent
materials, so we can monkey around with the EBO, running up high vertex
numbers without corresponding copies of the vertex data.

This patch as it stands is very broken and not in any way usable, and
much of it is basic hacky test code.